recall_old/requirements.txt
numpy
Pillow
rapidocr_onnxruntime
onnxruntime-gpu
nvidia-cudnn-cu12
pynvml
imagehash
pywin32
psutil
flask
requests
httpx
PySide6
chromadb
